fix: add checks
